Refactor the directory to match master branch
