+++ .env
@@ -1,1 +1,2 @@
-PORT=5007
+PORT=5004
+NAME=scene_detection
