+++ Dockerfile
@@ -4,6 +4,4 @@ WORKDIR /app
 
 COPY ./ /app
 
-RUN pip install --upgrade pip
-
 RUN pip --no-cache-dir install -r requirements.txt
