+++ SceneDetection/scene_detect_model.py
