+++ model.py
@@ -1,5 +1,5 @@
 import asyncio
-
+from PIL import Image
 from SceneDetection import scene_detect_model
 
 async def init():
@@ -14,8 +14,12 @@ async def init():
 def predict(image_file):
     """
     Interface method between model and server. This signature must not be
-    changed and your model must be able to predict given this input
+    changed and your model must be able to predict given a file-like object
+    with the image as an input.
     """
+
+    image = Image.open(image_file.name, mode='r')
+
     scene_detect_result = scene_detect_model.get_scene_attributes(image_file)
     return {
         "Category Confidence": scene_detect_result['category_results'],
