+++ requirements.txt
@@ -1,13 +1,13 @@
+pydantic
 fastapi
 pyinstaller
 setuptools
 uuid
 uvicorn
 python-multipart
+pillow
 scipy==1.1.0
 numpy==1.18.1
-Pillow==7.0.0
-redis
 tensorflow
 torch==1.4.0
 torchvision==0.5.0
