--- server/README.md
@@ -1,1 +1,0 @@
-# Server
